src/requirement.txt
antlr4-python3-runtime==4.9.3
beautifulsoup4==4.12.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
hydra-core==1.3.2
idna==3.4
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.0
pandas==2.0.0
polars==0.16.18
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.9
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
antlr4-python3-runtime==4.9.3
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
httplib2==0.22.0
hydra-core==1.3.2
idna==3.4
install==1.3.5
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
outcome==1.2.0
packaging==23.0
pandas==2.0.0
polars==0.16.18
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==12.1.3
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.9
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
webdriver-manager==3.8.5
wsproto==1.2.0
